♻️ Split up E2E utils
