+++ tests/fixtures.ts
@@ -3,7 +3,7 @@ import { test as base, type Locator } from '@playwright/test';
 
 import type { Settings } from '$lib/database';
 import type { IDBDatabaseType } from '$lib/idb.svelte';
-import type { RuntimeValue } from '$lib/metadata/index.js';
+import type { RuntimeValue } from '$lib/schemas/metadata';
 import type { ExportedProtocol } from '$lib/schemas/protocols';
 import type { Toast } from '$lib/toasts.svelte.js';
 import { safeJSONParse } from '$lib/utils';
@@ -31,7 +31,7 @@ import {
 	waitForLoadingEnd,
 	type NavigationTab,
 	type PredownloadedModel
-} from './utils.js';
+} from './utils/index.js';
 
 const fullProtocol = _fullProtocol as ExportedProtocol;
