+++ tests/import.spec.js
@@ -14,7 +14,7 @@ import {
 	newSession,
 	observationCard,
 	setInferenceModels
-} from './utils.js';
+} from './utils/index.js';
 
 test.describe('correct results.zip', () => {
 	test.beforeEach(async ({ page }) => {
