+++ tests/modal.spec.js
@@ -1,6 +1,6 @@
 import { issue } from './annotations.js';
 import { expect, test } from './fixtures.js';
-import { newSession } from './utils.js';
+import { newSession } from './utils/index.js';
 
 test.describe('closing a modal', () => {
 	test('when clicking outside', issue(469), async ({ page, app }) => {
